changed the return type to (int). cant synchro

+++ Minion/src/controller/LengthDistribution.java
@@ -34,7 +34,7 @@ public class LengthDistribution {
 			i++;
 		}
 		
-		return l.getLength(i);
+		return (int)l.getLength(i);
 	}
